Facebook Scraper/requirements.txt
# Core scraper dependencies
selenium>=4.15.0
webdriver-manager>=4.0.0
pandas>=2.0.0
openpyxl>=3.1.0
numpy>=1.24.0
python-dateutil>=2.8.0
feedparser>=6.0.0
requests>=2.31.0
python-dotenv>=1.0.0
python-Levenshtein>=0.21.0
praw>=7.7.0
beautifulsoup4>=4.12.0

# Database
psycopg2-binary>=2.9.0

# Celery for task scheduling
celery>=5.3.0
redis>=5.0.0
